requirements.txt
django
django-extensions
ipython
unipath
djangorestframework
